ConcludeInstigated is defined in the communication module
